Switch to using Sonatype OSS for releases.
This follows the guide on http://central.sonatype.org/pages/apache-maven.html
for how to update your POM to allow for deployment to Sonatype. To make a 
release you need to use the release profile (-Prelease argument for maven).

It also locks down the maven plugin versions and updates the reporting
section so that mvn site works.

Fixes: #2